+++ src/simmate/website/core_components/components/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .dynamic_form import DynamicFormComponent
-from .mixins import MoleculeInput, UserInput